+++ qiskit/transpiler/passes/routing/commuting_2q_gate_routing/swap_strategy.py
